Remove mock tests and focus on real extension testing
