+++ package.json
@@ -7,7 +7,6 @@
     "build": "echo 'No build step required'",
     "test": "playwright test",
     "test:debug": "playwright test --debug",
-    "test:real-world": "playwright test tests/real-world-sync.spec.js",
     "start:staging-peer": "cd staging-peer && npm start"
   },
   "dependencies": {
