--- tests/test-extension/manifest.json
@@ -1,29 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "HyperDB History Sync (Test)",
-  "version": "1.0",
-  "description": "Test version of HyperDB History Sync for integration testing",
-  "permissions": [
-    "history",
-    "storage",
-    "tabs",
-    "activeTab"
-  ],
-  "background": {
-    "service_worker": "background.js",
-    "type": "module"
-  },
-  "action": {
-    "default_popup": "popup.html",
-    "default_icon": {
-      "16": "../../images/icon16.png",
-      "48": "../../images/icon48.png",
-      "128": "../../images/icon128.png"
-    }
-  },
-  "icons": {
-    "16": "../../images/icon16.png",
-    "48": "../../images/icon48.png",
-    "128": "../../images/icon128.png"
-  }
-}
