--- tests/test-extension/popup.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>HyperDB History Sync</title>
-  <style>
-    body {
-      width: 400px;
-      padding: 10px;
-      font-family: Arial, sans-serif;
-    }
-    .status {
-      margin-top: 10px;
-      padding: 5px;
-      border-radius: 3px;
-    }
-    .connected {
-      background-color: #d4edda;
-      color: #155724;
-    }
-    .disconnected {
-      background-color: #f8d7da;
-      color: #721c24;
-    }
-    button {
-      margin-top: 10px;
-      padding: 5px 10px;
-      background-color: #007bff;
-      color: white;
-      border: none;
-      border-radius: 3px;
-      cursor: pointer;
-    }
-    button:hover {
-      background-color: #0069d9;
-    }
-    .peer-id {
-      word-break: break-all;
-      font-size: 12px;
-      margin-top: 10px;
-      padding: 5px;
-      background-color: #f8f9fa;
-      border-radius: 3px;
-    }
-    .history-list {
-      margin-top: 20px;
-      max-height: 300px;
-      overflow-y: auto;
-      border: 1px solid #ddd;
-      border-radius: 3px;
-      padding: 0;
-    }
-    .history-list li {
-      list-style: none;
-      padding: 10px;
-      border-bottom: 1px solid #eee;
-    }
-    .history-list li:last-child {
-      border-bottom: none;
-    }
-    .history-item .title {
-      font-weight: bold;
-    }
-    .history-item .url {
-      font-size: 12px;
-      color: #6c757d;
-      word-break: break-all;
-    }
-    .history-item .meta {
-      font-size: 11px;
-      color: #6c757d;
-      margin-top: 5px;
-    }
-    .section {
-      margin-top: 20px;
-      padding-top: 10px;
-      border-top: 1px solid #eee;
-    }
-  </style>
-</head>
-<body>
-  <h2>HyperDB History Sync</h2>
-  <div id="status" class="status disconnected">Disconnected</div>
-  <div id="peerId" class="peer-id"></div>
-  <div id="stagingPeerStatus" class="status disconnected">Staging Peer: Not Connected</div>
-  
-  <div class="section">
-    <button id="syncNow">Sync Now</button>
-    <button id="clearData">Clear Local Data</button>
-  </div>
-  
-  <div class="section">
-    <h3>History Items</h3>
-    <ul id="historyList" class="history-list">
-      <li>Loading history items...</li>
-    </ul>
-  </div>
-  
-  <script src="popup.js"></script>
-</body>
-</html>
